Se modifica componente brands a la misma estructura de techstack
