+++ components/Brands/constants.tsx
@@ -1,0 +1,60 @@
+import { Brand } from "@/types/brand";
+
+const brandsData: Brand[] = [
+    {
+      id: 1,
+      name: "IBM",
+      href: "https://www.ibm.com",
+      image: "/images/brands/ibm.svg",
+    },
+    {
+      id: 2,
+      name: "Televisa",
+      href: "https://www.televisa.com",
+      image: "/images/brands/Logo_de_Grupo_Televisa.svg",
+    },
+    {
+      id: 3,
+      name: "Univision",
+      href: "https://www.univision.com",
+      image: "/images/brands/Univision_logo.svg",
+    },
+    {
+      id: 4,
+      name: "Universal",
+      href: "https://www.universalmusic.com",
+      image: "/images/brands/universal-3.svg",
+    },
+    {
+      id: 5,
+      name: "Provectus",
+      href: "https://provectus.com",
+      image: "/images/brands/provectus.svg",
+    },
+    {
+      id: 6,
+      name: "Automation",
+      href: "https://www.automationanywhere.com",
+      image: "/images/brands/Automation.svg",
+    },
+    {
+      id: 7,
+      name: "Prodigious",
+      href: "https://www.prodigious.com",
+      image: "/images/brands/logo-prodigious.svg",
+    },
+    {
+      id: 7,
+      name: "MyBasePay",
+      href: "https://mybasepay.com",
+      image: "/images/brands/MyBasePay.svg",
+    },
+    {
+      id: 8,
+      name: "ChenMed",
+      href: "https://www.chenmed.com",
+      image: "/images/brands/chenMed.svg",
+    },
+  ];
+
+  export default brandsData;
